% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme.R
\name{add_attribute}
\alias{add_attribute}
\title{Add an Attribute to a Scheme}
\usage{
add_attribute(scheme, ...)
}
\arguments{
\item{scheme}{(\code{DLscheme} object) The scheme to add attributes to.}

\item{...}{attributes to add to the scheme, as key, value pairs.}
}
\value{
The updated scheme.
}
\description{
This function adds an attribute to an existing model scheme.
}
