% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_elements_flow.R
\name{add_inputs}
\alias{add_inputs}
\title{FUNCTION_TITLE}
\usage{
add_inputs(flow, inputs = list())
}
\arguments{
\item{flow}{(name) PARAM_DESCRIPTION}

\item{inputs}{(call) PARAM_DESCRIPTION, Default: list()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[igraph]{add_vertices}}
}
