% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trainable_model.R
\name{add_trainable_model}
\alias{add_trainable_model}
\title{FUNCTION_TITLE}
\usage{
add_trainable_model(flow, scheme = NULL, inputs, output, subset = NULL,
  verbose = TRUE)
}
\arguments{
\item{flow}{(name) PARAM_DESCRIPTION}

\item{scheme}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{inputs}{(call) PARAM_DESCRIPTION, Default: list()}

\item{output}{(name) PARAM_DESCRIPTION}

\item{subset}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{(logical) PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{

}
