% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_input.R
\name{analyze_input}
\alias{analyze_input}
\title{FUNCTION_TITLE}
\usage{
analyze_input(info = NULL, input)
}
\arguments{
\item{info}{(name) PARAM_DESCRIPTION}

\item{input_templates}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
