% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_output.R
\name{analyze_output}
\alias{analyze_output}
\title{FUNCTION_TITLE}
\usage{
analyze_output(info = NULL, output)
}
\arguments{
\item{info}{(name) PARAM_DESCRIPTION}

\item{output}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[neurobase]{readnii}}
 \code{\link[dplyr]{near}}
}
