% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_coef.R
\name{bce_dice_loss}
\alias{bce_dice_loss}
\title{Binary Cross Entropy and Dice Loss}
\usage{
bce_dice_loss(y_true, y_pred)
}
\arguments{
\item{y_true}{(vector) True classes}

\item{y_pred}{(vector) Predicted classes}
}
\value{
The Dice coefficient.
}
\description{
This function computes the Binary Cross Entropy and Dice Loss between 
a true and a predicted classifications.
}
