% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_categorical.R
\name{block_categorical}
\alias{block_categorical}
\title{Categorical Block}
\usage{
block_categorical(object, hidden_layers = NULL, num_classes = 2,
  units = 1, concatenate = FALSE)
}
\arguments{
\item{object}{(keras object) The object where to append the block.}

\item{hidden_layers}{(list) List of layers, with types and corresponding parameters.}

\item{num_classes}{(integer) The number of classes in the output of the block.}

\item{units}{(integer) Number of output units of the block.}
}
\value{
The object composed with the categorical block.
}
\description{
This function is used to append a multivalued block to a previously existing object.
}
\details{
A categorical block consists on \code{units} different paths, with the common input \code{object}, 
each path is composed by a set of hidden layers (all with the same architecture defined in the 
arguments of this function), and ended by a dense layer with \code{num_classes} units and "softmax"
activation.

The final output is the concatenation of these \code{units} independent paths.
}
