% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_multivalued.R
\name{block_multivalued}
\alias{block_multivalued}
\title{Multivalued Block}
\usage{
block_multivalued(object, hidden_layers = list(), num_values = 2,
  units = 1)
}
\arguments{
\item{object}{(keras object) The object where to append the block.}

\item{hidden_layers}{(list) List of layers, with types and corresponding parameters.}

\item{num_values}{(integer) The number of different possible values in the output of the block.}

\item{units}{(integer) Number of output units of the block.}

\item{hidden_activation}{(character) Activation function to use in the hidden layers of the block.}

\item{hidden_dropout}{(numeric in [0, 1]) The dropout to use in the hidden layers.}

\item{params}{(list) List of parameters to apply, if not listed in the previous ones.}
}
\value{
The object composed with the multivalued block.
}
\description{
This function is used to append a multivalued block to a previously existing object.
}
\details{
A multivalued block consists on \code{units} different paths, with the common input \code{object}, 
each path is composed by a set of hidden layers (all with the same architecture defined in the 
arguments of this function), and ended by a dense layer with \code{num_values} units and "sigmoid"
activation followed by another dense layer with one unit and "linear" activation.

The final output is the concatenation of these \code{units} independent paths.
}
