% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_path.R
\name{block_paths}
\alias{block_paths}
\title{Create Independent Paths}
\usage{
block_paths(object, hidden_layers = NULL, num_paths = 2,
  finalize_layers = list(), concatenate = FALSE)
}
\arguments{
\item{object}{(keras object) The object where to append the block.}

\item{hidden_layers}{(list) List of layers, with types and corresponding parameters.}

\item{num_paths}{(integer) Number of paths to generate with the same topology.}

\item{finalize_layers}{(list) List of layers, with types and corresponding parameters, to be added at the end of each path. Useful when called from \code{\link{block_categorical}}, \code{\link{block_regression}}, etc.}

\item{concatenate}{(logical) Default = \code{FALSE}. Do we have to concatenate all outputs?}
}
\value{
A list of keras objects (if \code{concatenate == FALSE}) or a single object (if \code{concatenate == TRUE}),
representing the \code{num_paths} independent paths.
}
\description{
This function appends a set of independent paths with the same architecture to a keras object.
}
