% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_regression.R
\name{block_regression}
\alias{block_regression}
\title{FUNCTION_TITLE}
\usage{
block_regression(object, hidden_layers = list(), units = 1,
  output_activation = "linear")
}
\arguments{
\item{object}{(name) PARAM_DESCRIPTION}

\item{hidden_layers}{(NULL) PARAM_DESCRIPTION, Default: list()}

\item{units}{(numeric) PARAM_DESCRIPTION, Default: 1}

\item{output_activation}{(character) PARAM_DESCRIPTION, Default: 'linear'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{

}
