% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_resnet.R
\name{block_resnet}
\alias{block_resnet}
\title{ResNet Block}
\usage{
block_resnet(object, hidden_layers = NULL, hidden_activation = "relu",
  hidden_dropout = 0, params = NULL)
}
\arguments{
\item{object}{(keras object) The object where to append the block.}

\item{hidden_layers}{(list) List of layers, with types and corresponding parameters.}

\item{hidden_activation}{(character) Activation function to use in the hidden layers of the block.}

\item{hidden_dropout}{(numeric in [0, 1]) The dropout to use in the hidden layers.}

\item{params}{(list) List of parameters to apply, if not listed in the previous ones.}
}
\value{
The object composed with the ResNet block.
}
\description{
This function is used to append a ResNet block to a previously existing object.
}
\details{
A ResNet block consists on a set of hidden layers, and ended by a dense layer with the same number
of units as output units in \code{object} and a final layer that adds \code{object} and the last layer.
}
