% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{clf}
\alias{clf}
\title{Continuous Learning of Features Block}
\usage{
clf(...)
}
\arguments{
\item{...}{arguments for the \code{\link{block_clf}} function.}
}
\value{
A list with \code{type} = "clf" and \code{params} with the rest of parameters that define the block.
}
\description{
This function is a wrapper to define a CLF block.
}
