% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizes.R
\name{compute_batch_size}
\alias{compute_batch_size}
\title{Compute Optimal Batch Size}
\usage{
compute_batch_size(model)
}
\arguments{
\item{model}{(\code{DLmodel} object) Model for whcih to compute the optimal batch_size}
}
\value{
The optimal batch_size
}
\description{
This function computes an estimate of the optimal batch size to train a model.
}
