% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_layers.R
\name{concatenate_layers}
\alias{concatenate_layers}
\title{FUNCTION_TITLE}
\usage{
concatenate_layers(layer_list)
}
\arguments{
\item{layer_list}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[keras]{layer_concatenate}}
}
