% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_bytes}
\alias{convert_to_bytes}
\title{Convert a Character String to Bytes}
\usage{
convert_to_bytes(x)
}
\arguments{
\item{x}{(character) The string to convert}
}
\value{
The number of bytes corresponding to the character expression: "1 K" is 1024, "1 M" is 1024 ^ 2, ...
}
\description{
This function converts a string in the format "number unit" to the corresponding number of bytes.
}
