% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{create_flow}
\alias{create_flow}
\title{FUNCTION_TITLE}
\usage{
create_flow(name = "", inputs = list())
}
\arguments{
\item{inputs}{(call) PARAM_DESCRIPTION, Default: list()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[igraph]{make_empty_graph}},\code{\link[igraph]{add_vertices}}
}
