% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generator.R
\name{create_generator}
\alias{create_generator}
\title{FUNCTION_TITLE}
\usage{
create_generator(model, x_files, y_files = NULL, mode = "sampling",
  target_windows_per_file = 1024)
}
\arguments{
\item{model}{(name) PARAM_DESCRIPTION}

\item{x_files}{(name) PARAM_DESCRIPTION}

\item{y_files}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{mode}{(call) PARAM_DESCRIPTION, Default: c("sampling", "all")}

\item{target_windows_per_file}{(numeric) PARAM_DESCRIPTION, Default: 1024}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[keras]{to_categorical}}
}
