% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_generator_from_config.R
\name{create_generator_from_config}
\alias{create_generator_from_config}
\title{FUNCTION_TITLE}
\usage{
create_generator_from_config(config, x_files, y_files = NULL,
  mode = c("sampling", "all"), num_windows = NULL, max_sub_epochs = 5)
}
\arguments{
\item{config}{(name) PARAM_DESCRIPTION}

\item{x_files}{(name) PARAM_DESCRIPTION}

\item{y_files}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{mode}{(call) PARAM_DESCRIPTION, Default: c("sampling", "all")}

\item{num_windows}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{max_sub_epochs}{(numeric) PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[keras]{to_categorical}}
}
