% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inference_function_from_config.R
\name{create_inference_function_from_config}
\alias{create_inference_function_from_config}
\title{Generic Inference Function for a Model Configuration}
\usage{
create_inference_function_from_config(object)
}
\arguments{
\item{config}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[keras]{to_categorical}}
}
