% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_from_config.R
\name{create_model_from_config}
\alias{create_model_from_config}
\title{Builds a DLmodel from a Given Configuration}
\usage{
create_model_from_config(config)
}
\arguments{
\item{config}{(\code{DLconfig} object) The configuration to use to build a model.}
}
\value{
The \code{DLmodel} object created according to the given configuration.
}
\description{
This function uses a previously defined configuration (\code{DLconfig} object) to create and compile
a keras model, stored, with the rest of metadata, in a \code{DLmodel} object.
}
