% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vol_layers.R
\name{create_vol_layers}
\alias{create_vol_layers}
\title{Create Volume Layers Definition from Pattern}
\usage{
create_vol_layers(info, pattern)
}
\arguments{
\item{info}{(\code{DLproblem} object) Problem defintion as given by \code{\link{get_problem_info}}}

\item{pattern}{(list, a set of layer definitions) The pattern of layers to replicate}
}
\value{
A list with as many replications of the pattern as inputs are defined in the \code{info} object.
}
\description{
This function replicates the given pattern of layers definition according to a problem definition.
}
