% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_encoder_functions.R
\name{decode}
\alias{decode}
\title{Decode Inputs using AutoEncoder}
\usage{
decode(model, inputs)
}
\arguments{
\item{model}{(\code{DLmodel}) The model to use}

\item{inputs}{(array) Inputs to apply the model on.}
}
\value{
The outputs given by the decoder of the model.
}
\description{
Decode Inputs using AutoEncoder
}
