% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{define_config}
\alias{define_config}
\title{Configuration Definition}
\usage{
define_config(...)
}
\arguments{
\item{...}{attributes to use}
}
\value{
The configuration for a model, of class \code{DLconfig}. A new model can be created by using 
\code{\link{create_model_from_config}} using the output of this function.
}
\description{
This function is used to define the configuration to create a model.
}
\details{
There are lots of parameters to tweak a model. Inspect the configuration file \code{config.properties} 
to know more.
}
