% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_last_layer.R
\name{define_last_layer}
\alias{define_last_layer}
\title{Define the Last Layer of a Model according to Problem Type}
\usage{
define_last_layer(info, units, only_convolutionals = FALSE,
  force_categorical = TRUE, multioutput = TRUE, ...)
}
\arguments{
\item{info}{(environment) Problem information as returned by \code{\link{get_problem_info}}.}

\item{units}{(integer) Number of output units}

\item{only_convolutionals}{(logical) If \code{TRUE}, just returns convolutional layers, otherwise dense layers, Default: FALSE}

\item{force_categorical}{(logical) When only 2 classes in a classification problem, one can use a \code{softsign} regression or a categorical layer, Default: TRUE}

\item{multioutput}{(logical) Split the output units, each one with a loss function, Default: TRUE}

\item{...}{arguments passed to other functions.}
}
\value{
A list with:
- \code{last_layer}, The definition of the last layer,
- \code{output_activation}, The output activation function,
- \code{loss}, The loss function(s) to use for each output,
- \code{remap_classes}, A mapping from the class numbers in the original images and 1:num_classes,
- \code{binarise}, A logical indicating if there are only 2 classes and not using a categorical output,
- \code{num_classes}, The number of classes, in a classification problem,
- \code{y_label}, The set of labels in the original output image in a classification problem,
- \code{multioutput}, A logical to indicate if the output is split.
}
\description{
This function uses the information about a problem to define
a possible last layer.
}
