% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dense}
\alias{dense}
\title{Dense Layer}
\usage{
dense(units, ...)
}
\arguments{
\item{units}{(integer) Number of units in the dense layer}

\item{...}{other arguments for the creation of a dense layer.}
}
\value{
A list with \code{type} = "dense" and \code{params} with the rest of parameters for the dense layer.
}
\description{
This function is a wrapper to define a dense layer, which is built in \code{\link{create_model_from_config}}.
}
