% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_coef.R
\name{dice_loss}
\alias{dice_loss}
\title{Dice Loss}
\usage{
dice_loss(y_true, y_pred, smooth = 1)
}
\arguments{
\item{y_true}{(vector) True classes}

\item{y_pred}{(vector) Predicted classes}

\item{smooth}{(real) Smoothness factor in the computation, Default: 1}
}
\value{
The Dice loss as \code{1 - Dice Coefficient}.
}
\description{
This function computes the Dice loss between a true and a predicted classifications.
}
