% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.execute}
\alias{DLflow.execute}
\title{FUNCTION_TITLE}
\arguments{
\item{inputs}{(call) PARAM_DESCRIPTION, Default: list()}

\item{desired_outputs}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{initialize_outputs}{(logical) PARAM_DESCRIPTION, Default: TRUE}

\item{mode}{(call) PARAM_DESCRIPTION, Default: c("debug", "faster", "medium", "slower")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
