% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.initialize}
\alias{DLflow.initialize}
\title{FUNCTION_TITLE}
\arguments{
\item{name}{(character) PARAM_DESCRIPTION, Default: ''}

\item{inputs}{(call) PARAM_DESCRIPTION, Default: list()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
