% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.save}
\alias{DLflow.save}
\title{FUNCTION_TITLE}
\arguments{
\item{path}{(call) PARAM_DESCRIPTION, Default: tempdir()}

\item{file_prefix}{(call) PARAM_DESCRIPTION, Default: self$name()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
