% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.subset}
\alias{DLflow.subset}
\title{FUNCTION_TITLE}
\arguments{
\item{outputs}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
