% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLflow.R
\name{DLflow.train}
\alias{DLflow.train}
\title{FUNCTION_TITLE}
\arguments{
\item{output}{(name) PARAM_DESCRIPTION}

\item{input_filenames}{(name) PARAM_DESCRIPTION}

\item{output_filenames}{(name) PARAM_DESCRIPTION}

\item{train_split}{(numeric) PARAM_DESCRIPTION, Default: 0.75}

\item{epochs}{(numeric) PARAM_DESCRIPTION, Default: 10}

\item{target_windows_per_file}{(numeric) PARAM_DESCRIPTION, Default: 1024}

\item{mode}{(call) PARAM_DESCRIPTION, Default: c("debug", "faster", "medium", "slower")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
