% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLmodel.R
\name{DLmodel.add_to_history}
\alias{DLmodel.add_to_history}
\title{FUNCTION_TITLE}
\arguments{
\item{epoch}{(numeric) PARAM_DESCRIPTION, Default: 0}

\item{subepoch}{(numeric) PARAM_DESCRIPTION, Default: 0}

\item{time}{(call) PARAM_DESCRIPTION, Default: Sys.time()}

\item{loss}{(logical) PARAM_DESCRIPTION, Default: NA}

\item{val_loss}{(logical) PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
