% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLmodel.R
\name{DLmodel.check_memory}
\alias{DLmodel.check_memory}
\title{FUNCTION_TITLE}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[prettyunits]{pretty_bytes}}
}
