% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLmodel.R
\name{DLmodel.fit}
\alias{DLmodel.fit}
\title{FUNCTION_TITLE}
\arguments{
\item{epochs}{(numeric) PARAM_DESCRIPTION, Default: 10}

\item{keep_best}{(logical) PARAM_DESCRIPTION, Default: TRUE}

\item{metrics_viewer}{(logical) PARAM_DESCRIPTION, Default: FALSE}

\item{...}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
