% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLmodel.R
\name{DLmodel.infer}
\alias{DLmodel.infer}
\title{FUNCTION_TITLE}
\arguments{
\item{V}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{speed}{(call) PARAM_DESCRIPTION, Default: c("faster", "medium", "slower")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
