% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLmodel.R
\name{DLmodel.log}
\alias{DLmodel.log}
\title{FUNCTION_TITLE}
\arguments{
\item{level}{(call) PARAM_DESCRIPTION, Default: c("DEBUG", "INFO", "WARNING", "ERROR")}

\item{message}{(character) PARAM_DESCRIPTION, Default: '...'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
