% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLmodel.R
\name{DLmodel.print_log}
\alias{DLmodel.print_log}
\title{FUNCTION_TITLE}
\arguments{
\item{level}{(call) PARAM_DESCRIPTION, Default: c("DEBUG", "WARNING", "INFO", "ERROR")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
