% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLmodel.R
\docType{class}
\name{DLmodel}
\alias{DLmodel}
\title{DLmodel Class}
\format{\code{\link{R6Class}} object.}
\usage{
DLmodel
}
\value{
Object of \code{\link{R6Class}} and \code{DLmodel}.
}
\description{
DLmodel Class
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method follow the corresponding link. }
  \item{\code{initialize(...)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.initialize}.}
  \item{\code{update(...)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.update}.}
  \item{\code{summary()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.summary}.}
  \item{\code{log(level = c("DEBUG", "INFO", "WARNING", "ERROR"), message = '...')}}{METHOD DESCRIPTION. Documented in \link{DLmodel.log}.}
  \item{\code{add_to_history(epoch = 0, subepoch = 0, time = Sys.time(), loss = NA, val_loss = NA)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.add_to_history}.}
  \item{\code{plot_history()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.plot_history}.}
  \item{\code{reset_history()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.reset_history}.}
  \item{\code{render_history(initialize = TRUE)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.render_history}.}
  \item{\code{update_render()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.update_render}.}
  \item{\code{get_model()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.get_model}.}
  \item{\code{get_width()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.get_width}.}
  \item{\code{set_width(width)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.set_width}.}
  \item{\code{get_loss()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.get_loss}.}
  \item{\code{set_loss(loss)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.set_loss}.}
  \item{\code{get_encoder()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.get_encoder}.}
  \item{\code{get_decoder()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.get_decoder}.}
  \item{\code{get_config()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.get_config}.}
  \item{\code{get_history()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.get_history}.}
  \item{\code{print_log(level = c("DEBUG", "WARNING", "INFO", "ERROR"))}}{METHOD DESCRIPTION. Documented in \link{DLmodel.print_log}.}
  \item{\code{errors()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.errors}.}
  \item{\code{warnings()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.warnings}.}
  \item{\code{save_log(filename, level = c("DEBUG", "WARNING", "INFO", "ERROR"))}}{METHOD DESCRIPTION. Documented in \link{DLmodel.save_log}.}
  \item{\code{check_memory()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.check_memory}.}
  \item{\code{graph()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.graph}.}
  \item{\code{plot(to_file)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.plot}.}
  \item{\code{load(path, prefix)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.load}.}
  \item{\code{save(path, prefix, comment)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.save}.}
  \item{\code{use_data(use = c("train", "test"), x_files, y_files = NULL, target_windows_per_file = 1024)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.use_data}.}
  \item{\code{fit(epochs = 10, keep_best = TRUE, metrics_viewer = FALSE, ...)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.fit}.}
  \item{\code{reset()}}{METHOD DESCRIPTION. Documented in \link{DLmodel.reset}.}
  \item{\code{infer(V = NULL, speed = c("faster", "medium", "slower"))}}{METHOD DESCRIPTION. Documented in \link{DLmodel.infer}.}
  \item{\code{clone(deep = FALSE)}}{METHOD DESCRIPTION. Documented in \link{DLmodel.clone}.}
 }
}

\examples{
DLmodel$new()

}
\keyword{data}
