% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLmodel.R
\name{DLmodel.save}
\alias{DLmodel.save}
\title{FUNCTION_TITLE}
\arguments{
\item{path}{(name) PARAM_DESCRIPTION}

\item{prefix}{(name) PARAM_DESCRIPTION}

\item{comment}{(character) PARAM_DESCRIPTION, Default: ''}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
