% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLmodel.R
\name{DLmodel.set_width}
\alias{DLmodel.set_width}
\title{FUNCTION_TITLE}
\arguments{
\item{width}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
