% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLmodel.R
\name{DLmodel.update}
\alias{DLmodel.update}
\title{FUNCTION_TITLE}
\arguments{
\item{...}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
