% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLmodel.R
\name{DLmodel.use_data}
\alias{DLmodel.use_data}
\title{FUNCTION_TITLE}
\arguments{
\item{use}{(call) PARAM_DESCRIPTION, Default: c("train", "test")}

\item{x_files}{(name) PARAM_DESCRIPTION}

\item{y_files}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{target_windows_per_file}{(numeric) PARAM_DESCRIPTION, Default: 1024}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
