% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLscheme.R
\name{DLscheme.clone}
\alias{DLscheme.clone}
\title{FUNCTION_TITLE}
\arguments{
\item{deep}{(logical) PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
