% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_DLscheme.R
\name{DLscheme.from_model}
\alias{DLscheme.from_model}
\title{FUNCTION_TITLE}
\arguments{
\item{model}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
