% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{double}
\alias{double}
\title{Doubling Convolutional Block}
\usage{
double(initial_filters = 2, kernel_size = c(3, 3, 3), ...)
}
\arguments{
\item{initial_filters}{(numeric) Number of filters in the first convolutional layer, Default: 2}

\item{kernel_size}{(list or vector) size of the convolution kernels, Default: c(3, 3, 3)}

\item{...}{arguments for the \code{\link{block_double}} function.}
}
\value{
A list with \code{type} = "double" and \code{params} with the rest of parameters that define the block.
}
\description{
This function is a wrapper to define a doubling convolutional block.
}
