% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_encoder_functions.R
\name{encode}
\alias{encode}
\title{Encode Inputs using AutoEncoder}
\usage{
encode(model, inputs)
}
\arguments{
\item{model}{(\code{DLmodel}) The model to use}

\item{inputs}{(array) Inputs to apply the model on.}
}
\value{
The outputs given by the encoder of the model.
}
\description{
Encode Inputs using AutoEncoder
}
