% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_optimizer.R
\name{eval_optimizer}
\alias{eval_optimizer}
\title{Evaluate a Parsed Optimizer}
\usage{
eval_optimizer(parsed_optimizer)
}
\arguments{
\item{parsed_optimizer}{(list) The result of a call to \code{\link{parse_optimizer}}.}
}
\value{
A keras optimizer instance for the current session.
}
\description{
This function is the complement for \code{\link{parse_optimizer}}, since it reconstructs the original
optimizer from its configuration, as needed to compile a model.
}
