% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_flow.R
\name{execute_flow}
\alias{execute_flow}
\title{FUNCTION_TITLE}
\usage{
execute_flow(flow, inputs = list(), given_inputs = NULL,
  desired_outputs = NULL, initialize_outputs = TRUE, mode = c("debug",
  "faster", "medium", "slower"))
}
\arguments{
\item{flow}{(name) PARAM_DESCRIPTION}

\item{inputs}{(call) PARAM_DESCRIPTION, Default: list()}

\item{given_inputs}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{desired_outputs}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{initialize_outputs}{(logical) PARAM_DESCRIPTION, Default: TRUE}

\item{mode}{(call) PARAM_DESCRIPTION, Default: c("debug", "faster", "medium", "slower")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[neurobase]{readnii}}
}
