% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_with_generator.R
\name{fit_with_generator}
\alias{fit_with_generator}
\title{Fit a Model using Generators}
\usage{
fit_with_generator(.model, generator = NULL, steps_per_epoch = NULL,
  train_config = NULL, epochs = 10, starting_epoch = 1,
  validation_data = NULL, validation_steps = NULL,
  validation_config = NULL, keep_best = TRUE, verbose = TRUE,
  metrics_viewer = FALSE, reset_optimizer = FALSE, ...)
}
\arguments{
\item{.model}{(\code{DLmodel}) The model to train}

\item{generator}{(function) The generator function for training data, built by \code{\link{create_generator_from_config}}, Default: NULL}

\item{steps_per_epoch}{(integer) Number of steps per training epoch, Default: NULL}

\item{train_config}{(list) The training configuration, the output of \code{\link{create_generator_from_config}} when inputs are training files, Default: NULL}

\item{epochs}{(numeric) Maximum number of epochs to train, Default: 10}

\item{starting_epoch}{(numeric) strating epoch, useful when we want to resume a previous fit, Default: 1}

\item{validation_data}{(function or matrix) Data for validation. It can be a generator function, built by \code{\link{create_generator_from_config}}, Default: NULL}

\item{validation_steps}{(integer) Number of steps of validation per epoch, Default: NULL}

\item{validation_config}{(list) The testing configuration, the output of \code{\link{create_generator_from_config}} when inputs are validation files, Default: NULL}

\item{keep_best}{(logical) Should the training always store the best model up-to-date?, Default: TRUE}

\item{verbose}{(logical) Provide additional information on training, Default: TRUE}

\item{metrics_viewer}{(logical) Visualize training loss interactively while fitting?, Default: FALSE}

\item{reset_optimizer}{(logical) Reset optimizer state after each subepoch?, Default: FALSE}

\item{...}{extra arguments passed to other functions.}
}
\value{
The trained \code{DLmodel}.
}
\description{
this function is used to fit or train a \code{DLmodel} by using generator 
functions for training and validation data.
}
\details{
\code{generator}, \code{steps_per_epoch}, \code{validation_data} and \code{validation_steps} are completely and automatically determined if one uses the \code{train_config} and \code{validation_config} parameters, both being the outputs of \code{\link{create_generator_from_config}}.

Additionally, we can pass this function two arguments: \code{path} and \code{prefix}, the best model will be stored in the corresponding path with the given prefix (usually something indicative or with a timestamp).
}
\seealso{
\code{\link[keras]{callback_lambda}}
}
