% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_by_gaussian_kernel.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{FUNCTION_TITLE}
\usage{
gaussian_kernel(sigma = 1, dim = 3, size = 3, normalised = TRUE)
}
\arguments{
\item{sigma}{(numeric) PARAM_DESCRIPTION, Default: 1}

\item{dim}{(numeric) PARAM_DESCRIPTION, Default: 3}

\item{size}{(numeric) PARAM_DESCRIPTION, Default: 3}

\item{normalised}{(logical) PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
