% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activations.R
\name{get_activations}
\alias{get_activations}
\title{Get Activations in Intermediate Layer}
\usage{
get_activations(object, layer)
}
\arguments{
\item{object}{(\code{DLmodel} object) Base model.}

\item{layer}{(integer or character) Name or index of the intermediate layer to compute its output.}
}
\value{
A function that computes the output of the layer given an input.
}
\description{
This function computes the activation of an intermediate layer of a model.
}
