% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dl4ni_config.R
\name{get_dl4ni_config}
\alias{get_dl4ni_config}
\title{Get Default Configuration}
\usage{
get_dl4ni_config()
}
\value{
An object of class \code{DLconfig} containing the configuration.
}
\description{
This function returns the default configuration stored in the \code{config.properties} file.
}
