% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_problem_info.R
\name{get_problem_info}
\alias{get_problem_info}
\title{FUNCTION_TITLE}
\usage{
get_problem_info(problem = "foo", input_path = NULL, output_path = NULL,
  num_subjects = NULL)
}
\arguments{
\item{problem}{(character) PARAM_DESCRIPTION, Default: 'foo'}

\item{input_path}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{output_path}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{num_subjects}{(NULL) PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[neurobase]{readnii}}
 \code{\link[dplyr]{near}}
 \code{\link[utils]{select.list}}
}
