% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_model.R
\name{graph_from_model}
\alias{graph_from_model}
\title{Create Graph from a Model}
\usage{
graph_from_model(.model, mode = c("graphviz", "igraph"))
}
\arguments{
\item{.model}{(\code{DLmodel}) Model for which to build the graph.}
}
\value{
An \code{igraph} object.
}
\description{
This function creates a graph representation for a given model.
}
\details{
DETAILS
}
\seealso{
\code{\link[igraph]{graph.adjacency}}
 \code{\link[jsonlite]{fromJSON}}
 \code{\link[keras]{model_to_json}}
}
