% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{layer_units}
\alias{layer_units}
\title{Layer Output Units}
\usage{
layer_units(object)
}
\arguments{
\item{object}{(\code{keras} object) The object to which compute shape.}
}
\value{
A vector with the \code{object} shape
}
\description{
This function gets the total number of units of the output of a layer in a model.
}
