% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_flow.R
\name{load_flow}
\alias{load_flow}
\title{FUNCTION_TITLE}
\usage{
load_flow(filename, verbose = TRUE)
}
\arguments{
\item{filename}{(name) PARAM_DESCRIPTION}

\item{verbose}{(logical) PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
