% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{load_model}
\alias{load_model}
\title{Load DLmodel}
\usage{
load_model(path, prefix)
}
\arguments{
\item{path}{(character) Path where the model is stored.}

\item{prefix}{(character) Subfolder of the \code{path} where all files are stored.}
}
\value{
A \code{DLmodel} object.
}
\description{
This function loads different files conforming a \code{DLmodel} object.
}
\seealso{
\code{\link[keras]{model_from_json}},\code{\link[keras]{load_model_weights_hdf5}},
 \code{\link{load_model}}
}
