% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ids.R
\name{map_ids}
\alias{map_ids}
\title{FUNCTION_TITLE}
\usage{
map_ids(image, remap_classes, invert = FALSE)
}
\arguments{
\item{image}{(name) PARAM_DESCRIPTION}

\item{remap_classes}{(name) PARAM_DESCRIPTION}

\item{invert}{(logical) PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
