% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizes.R
\name{model_size}
\alias{model_size}
\title{Compute Model Total Size}
\usage{
model_size(model)
}
\arguments{
\item{model}{(\code{DLmodel} object) Model for whcih to compute the optimal batch_size}
}
\value{
Total model size in bytes
}
\description{
This function computes the total size of a model, in bytes, for trainable parameters
and the input data.
}
