% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{multivalued}
\alias{multivalued}
\title{Multivalued Block}
\usage{
multivalued(num_values, units, ...)
}
\arguments{
\item{num_values}{(integer) Number of values for the block}

\item{units}{(integer) Number of units in the output block}

\item{...}{other arguments for the creation of a multivalued block.}
}
\value{
A list with \code{type} = "multivalued" and \code{params} with the rest of parameters for the categorical block.
}
\description{
This function is a wrapper to define a multivalued block, which is built in \code{\link{create_model_from_config}}.
}
