% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_layers.R
\name{normalize_layers}
\alias{normalize_layers}
\title{Normalize Layers}
\usage{
normalize_layers(layers_definition)
}
\arguments{
\item{layers_definition}{(list) List of layers.}
}
\value{
The normalized list of layers.
}
\description{
This function takes as input a list of layers and checks that all are well-defined.
}
