% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{object_shape}
\alias{object_shape}
\title{Object Output Shape}
\usage{
object_shape(object)
}
\arguments{
\item{object}{(\code{keras} object) The object to which compute shape.}
}
\value{
A vector with the \code{object} shape
}
\description{
This function gets the shape of the output of a layer in a model.
}
