% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_plot.R
\name{ortho_plot}
\alias{ortho_plot}
\title{Orthographic Plot}
\usage{
ortho_plot(x, y = NULL, text = "", interactiveness = FALSE,
  force = FALSE, ...)
}
\arguments{
\item{x}{(\code{nifti} image) Image to plot}

\item{y}{(\code{nifti} image) If not NULL, it is an overlay, Default: NULL}

\item{text}{(character) Text used as title of the plot, Default: NULL}

\item{interactiveness}{(logical) If \code{TRUE}, use package \code{papayar} to render an htmlwidget. If \code{FALSE}, use \code{neurobase::ortho2}, Default: TRUE}

\item{force}{(logical) Force saving of plot (forces \code{interactiveness = FALSE}), Default: FALSE}

\item{...}{other arguments, such as the \code{saving_path} and \code{saving_prefix} to store in case of \code{force == TRUE}.}
}
\description{
This function renders an orthographic plot of a nifti image.
}
\seealso{
\code{\link[neurobase]{ortho2}}
 \code{\link[RNifti]{updateNifti}}
 \code{\link[RNifti]{updateNifti}}
}
