% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_optimizer.R
\name{parse_optimizer}
\alias{parse_optimizer}
\title{Parse Keras Optimizer}
\usage{
parse_optimizer(optimizer)
}
\arguments{
\item{optimizer}{(character or keras optimizer) Optimizer name (e.g. "adadelta") or an instance of a
optimizer (e.g. optimizer_adadelta()).}
}
\value{
A list with the name of the optimizer and its configuration.
}
\description{
This function parses the configuration of the optimizer used to train a model 
in order to be able to save and load it.
}
