% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{plot_flow}
\alias{plot_flow}
\title{FUNCTION_TITLE}
\usage{
plot_flow(flow, interactive = FALSE)
}
\arguments{
\item{flow}{(name) PARAM_DESCRIPTION}

\item{interactive}{(logical) PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[threejs]{graphjs}}
 \code{\link[igraph]{layout_with_sugiyama}}
 \code{\link[scales]{alpha}},\code{\link[scales]{hue_pal}}
}
