% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_model.R
\name{plot_graph}
\alias{plot_graph}
\title{FUNCTION_TITLE}
\usage{
plot_graph(g, interactive = FALSE)
}
\arguments{
\item{g}{(name) PARAM_DESCRIPTION}

\item{interactive}{(logical) PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[igraph]{layout_with_sugiyama}}
 \code{\link[scales]{alpha}},\code{\link[scales]{hue_pal}}
}
