% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_files_for_inference.R
\name{prepare_files_for_inference}
\alias{prepare_files_for_inference}
\title{Read Multiple Files}
\usage{
prepare_files_for_inference(file_list)
}
\arguments{
\item{file_list}{(list) Path of files to import}
}
\value{
A list with one entry for each file read.
}
\description{
This function reads several NIfTI files.
}
