% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nifti_to_array.R
\name{read_nifti_to_array}
\alias{read_nifti_to_array}
\title{Read Nifti File}
\usage{
read_nifti_to_array(filename)
}
\arguments{
\item{filename}{(character) The path to the file to be read}
}
\value{
An array with the NIfTI image.
}
\description{
This function is used to read a NIfTI file and convert the object into an array
}
