% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{regression}
\alias{regression}
\title{Regression Block}
\usage{
regression(units, ...)
}
\arguments{
\item{units}{(integer) Number of units in the output layer.}

\item{...}{other arguments for the creation of a regression block.}
}
\value{
A list with \code{type} = "regression" and \code{params} with the rest of parameters for the output block.
}
\description{
This function is a wrapper to define a regression block, which is built in \code{\link{create_model_from_config}}.
}
