% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{resnet}
\alias{resnet}
\title{ResNet Block}
\usage{
resnet(...)
}
\arguments{
\item{...}{arguments for the \code{\link{block_resnet}} function.}
}
\value{
A list with \code{type} = "resnet" and \code{params} with the rest of parameters that define the block.
}
\description{
This function is a wrapper to define a ResNet block.
}
