% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load_flow.R
\name{save_flow}
\alias{save_flow}
\title{FUNCTION_TITLE}
\usage{
save_flow(flow, path = tempdir(), file_prefix = flow$name)
}
\arguments{
\item{flow}{(name) PARAM_DESCRIPTION}

\item{path}{(call) PARAM_DESCRIPTION, Default: tempdir()}

\item{file_prefix}{(call) PARAM_DESCRIPTION, Default: flow$name}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[zip]{zip}}
}
