% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_load.R
\name{save_model}
\alias{save_model}
\title{Save DLmodel}
\usage{
save_model(.model, path, prefix, comment = "")
}
\arguments{
\item{.model}{(\code{DLmodel}) Model to save.}

\item{path}{(character) Path where to save the model.}

\item{prefix}{(character) Subfolder of the \code{path} where to store all files}

\item{comment}{(character) Comment to include together with the model parameters, Default: ''}
}
\description{
This function saves all data related to a model in a given path.
}
