% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segnet.R
\name{segnet}
\alias{segnet}
\title{imlab-uiip/keras-segnet}
\usage{
segnet(initial_filters = 2, kernel_size = c(3, 3, 3), depth = 5,
  mode = c("sampling", "convolutional"))
}
\description{
imlab-uiip/keras-segnet
}
