% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_trainability.R
\name{set_trainability}
\alias{set_trainability}
\title{Set Trainability of Layers in Model}
\usage{
set_trainability(.model, layer_names = NULL, trainability = FALSE)
}
\arguments{
\item{layer_names}{(character vector) Names of the layers to set their \code{trainable} attribute.}

\item{trainability}{(logical vector) New values for the \code{trainable} attribute of the selected layers.}

\item{model}{(\code{DLmodel}) The model built by \code{\link{create_model_from_config}}.}
}
\value{
changes the \code{trainable} attribute in the selected layers in the model, in place.
}
\description{
This function sets the \code{trainable} attribute of a subset of the layers
in a model to \code{TRUE} or \code{FALSE}.
}
