% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_by_gaussian_kernel.R
\name{smooth_by_gaussian_kernel}
\alias{smooth_by_gaussian_kernel}
\title{FUNCTION_TITLE}
\usage{
smooth_by_gaussian_kernel(image, kernel_sigma, kernel_width)
}
\arguments{
\item{image}{(name) PARAM_DESCRIPTION}

\item{kernel_sigma}{(name) PARAM_DESCRIPTION}

\item{kernel_width}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
