% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_train_test_sets.R
\name{split_train_test_sets}
\alias{split_train_test_sets}
\title{Split Problem into Train and Test Sets}
\usage{
split_train_test_sets(info, train_split = 0.75)
}
\arguments{
\item{info}{(\code{DLproblem} object, as created by \code{\link{get_rpoblem_info}}) The information of the problem, with inputs and outputs.}

\item{train_split}{(numeric) Fraction of the subjects to be used as training set, Default: 0.75}
}
\value{
The \code{DLproblem} completed with \code{train} and \code{test} sets.
}
\description{
This function is used to split inputs and outputs of a problem definition into train and test sets.
}
