% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_flow.R
\name{subset_flow}
\alias{subset_flow}
\title{FUNCTION_TITLE}
\usage{
subset_flow(flow, outputs)
}
\arguments{
\item{flow}{(name) PARAM_DESCRIPTION}

\item{outputs}{(name) PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[igraph]{topo_sort}},\code{\link[igraph]{make_empty_graph}},\code{\link[igraph]{add_vertices}},\code{\link[igraph]{add_edges}}
}
