% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_problem.R
\name{subset_problem}
\alias{subset_problem}
\title{FUNCTION_TITLE}
\usage{
subset_problem(problem_info, subset_classes, unify_classes = NULL,
  use_all = TRUE)
}
\arguments{
\item{problem_info}{(name) PARAM_DESCRIPTION}

\item{subset_classes}{(name) PARAM_DESCRIPTION}

\item{unify_classes}{(name) PARAM_DESCRIPTION}

\item{use_all}{(logical) PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
