% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_categorical_volume.R
\name{to_categorical_volume}
\alias{to_categorical_volume}
\title{Split a Label Volume in One-hot Encoding}
\usage{
to_categorical_volume(V, unique_labels = sort(unique(as.vector(V))))
}
\arguments{
\item{V}{(3D array) Input volume}
}
\value{
A 4D volume representing the one-hot encoding of the input 3D volume.
}
\description{
This function splits a volume (representing a labelling, such as the result of a segmentation) into several binary volumes (concetenated as a 4D volume) such that each individual 3D volume indicates the presence of a class in the image.
}
