% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_flow.R
\name{train_output}
\alias{train_output}
\title{FUNCTION_TITLE}
\usage{
train_output(flow, output, input_filenames, output_filenames,
  given_input = NULL, train_split = 0.75, epochs = 10,
  target_windows_per_file = 1024, metrics_viewer = FALSE,
  mode = c("debug", "faster", "medium", "slower"), verbose = TRUE)
}
\arguments{
\item{flow}{(name) PARAM_DESCRIPTION}

\item{output}{(name) PARAM_DESCRIPTION}

\item{input_filenames}{(name) PARAM_DESCRIPTION}

\item{output_filenames}{(name) PARAM_DESCRIPTION}

\item{given_input}{(NULL) PARAM_DESCRIPTION, Default: NULL}

\item{train_split}{(numeric) PARAM_DESCRIPTION, Default: 0.75}

\item{epochs}{(numeric) PARAM_DESCRIPTION, Default: 10}

\item{mode}{(call) PARAM_DESCRIPTION, Default: c("debug", "faster", "medium", "slower")}

\item{verbose}{(logical) PARAM_DESCRIPTION, Default: TRUE}

\item{max_sub_epochs}{(numeric) PARAM_DESCRIPTION, Default: 5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[neurobase]{fast_readnii}}
 \code{\link[oro.nifti]{as.nifti}}
 \code{\link[RNifti]{writeNifti}}
 \code{\link[oro.nifti]{as.nifti}}
 \code{\link[RNifti]{writeNifti}}
}
