% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_coords.R
\name{transform_coords}
\alias{transform_coords}
\title{Tranform Coords Between Images}
\usage{
transform_coords(x, y, z, Vx, Vy)
}
\arguments{
\item{x}{(numeric vector) X coordinates in original image.}

\item{y}{(numeric vector) Y coordinates in original image.}

\item{z}{(numeric vector) Z coordinates in original image.}

\item{Vx}{(3d array) Original image.}

\item{Vy}{(3d array) Target image.}
}
\value{
The transformed \code{x}, \code{y} and \code{z}.
}
\description{
This function transform coordinates between images with, possibly, different dimensions.
}
